Move wrappers logic into reify/node/wrappers.js.
